STADTPULS-383: chore(housekeeping): split test utils
